+++ src/com/kaylerrenslow/armaDialogCreator/main/HelpUrls.java
@@ -7,9 +7,12 @@ public interface HelpUrls {
 
 	String WIKI_URL = "https://github.com/kayler-renslow/arma-dialog-creator/wiki";
 	String MACROS = "https://github.com/kayler-renslow/arma-dialog-creator/wiki/Macros";
-	String PROJECT_INIT_WINDOW = "https://github.com/kayler-renslow/arma-dialog-creator/wiki/Create,-Open,-and-Export-Projects";
+	String PROJECT_INIT_WINDOW = "https://github.com/kayler-renslow/arma-dialog-creator/wiki/Create,-Open,-and-Import-a-Project";
 	String CONFIGURE_DIRECTORIES_POPUP = "https://github.com/kayler-renslow/arma-dialog-creator/wiki/Configuring-Directories";
-	String NEW_CUSTOM_CONTROL_POPUP = "https://github.com/kayler-renslow/arma-dialog-creator/wiki/Creating-a-Custom-Control";
-	String EXPORT = "https://github.com/kayler-renslow/arma-dialog-creator/wiki/Create,-Open,-and-Export-Projects#exporting-a-project";
+	String CUSTOM_CONTROL_CLASSES = "https://github.com/kayler-renslow/arma-dialog-creator/wiki/Custom-Control-Classes";
+	String EXPORT = "https://github.com/kayler-renslow/arma-dialog-creator/wiki/Project-Exporting";
 	String NEW_CONTROL_POPUP = "https://github.com/kayler-renslow/arma-dialog-creator/wiki/Creating-a-New-Control";
+	String STRING_TABLE_EDITOR = "https://github.com/kayler-renslow/arma-dialog-creator/wiki/String-Table-Editor";
+	String MINI_SQF_EVALUATOR = "https://github.com/kayler-renslow/arma-dialog-creator/wiki/Mini-SQF-Evaluator";
+	String ARMA_PREVIEW_WINDOW = "https://github.com/kayler-renslow/arma-dialog-creator/wiki/Arma-Preview-Window";
 }